simplified the sequencers init flow
